+++ Assets/WorldGeneration/_Scripts/ScriptableObjects/BurrowSettings.cs
@@ -1,0 +1,14 @@
+using System;
+using Time;
+using UnityEngine;
+
+namespace WorldGeneration._Scripts.ScriptableObjects
+{
+    [CreateAssetMenu(menuName = "ScriptableObjects/Settings/Burrow")]
+    [Serializable]
+    public class BurrowSettings : ScriptableObject
+    {
+        public InGameDate lifespan;
+        [HideInInspector] public bool isOccupied;
+    }
+}