+++ Assets/WorldGeneration/_Scripts/ScriptableObjects/PlantSettings.cs
@@ -1,0 +1,11 @@
+using Time;
+using UnityEngine;
+
+namespace WorldGeneration._Scripts.ScriptableObjects
+{
+    [CreateAssetMenu(menuName = "ScriptableObjects/Settings/Plant")]
+    public class PlantSettings : ScriptableObject
+    {
+        public InGameDate lifespan;
+    }
+}