+++ Assets/WorldGeneration/_Scripts/Spawning/BurrowSpawner/RabbitBurrowSpawner.cs
@@ -1,0 +1,16 @@
+using UnityEngine;
+
+namespace WorldGeneration._Scripts.Spawning.BurrowSpawner
+{
+    public class RabbitBurrowSpawner : BurrowSpawner
+    {
+        public override void Spawn(GameObject burrow, Vector3 burrowPos)
+        {
+            // float xSize = mesh.
+            // Vector3 newPos = new Vector3(Random.Range(xMin, xMax), Random.Range(yMin, yMax), 0);
+            // GameObject octo = Instantiate(enemyPrefab, newPos, Quaternion.identity) as GameObject;
+            
+            Object.Instantiate(burrow, burrowPos, Quaternion.identity);
+        }
+    }
+}