+++ Assets/WorldGeneration/_Scripts/TerrainGeneration/NoiseWithClamp.cs
@@ -1,0 +1,10 @@
+using WorldGeneration._Scripts.Helper;
+
+namespace WorldGeneration._Scripts.TerrainGeneration
+{
+    public struct NoiseWithClamp
+    {
+        public NoiseGenerator NoiseGenerator;
+        public ValueClamp ValueClamp;
+    }
+}